+++ src/main/java/com/ldb/truck/Entity/OrderItem/V_order_item_details.java
@@ -39,10 +39,10 @@ public class V_order_item_details {
     private Integer exchangeRate;
 
     @Column(name = "qty")
-    private Integer qty;
+    private BigDecimal  qty;
 
     @Column(name = "price")
-    private Float price;
+    private BigDecimal  price;
 
     @Column(name = "status", length = 200)
     private String status;
