+++ src/main/java/com/ldb/truck/Entity/RequestItem/RequestItem.java
@@ -5,13 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 public class RequestItem {
     private int item;
-    private int qty;
+    private BigDecimal qty;
     private String type;
     private String borNo;
     private String mchNo;
